project-backend-mains/project-backend-main/requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-dotenv>=0.19.0
passlib>=1.7.4
python-jose[cryptography]>=3.3.0
sqlalchemy>=1.4.0
pymysql>=1.0.2
python-multipart>=0.0.5